\name{FilmsGmooG}
\alias{FilmsGmooG-package}
\docType{package}
\title{
FilmsGmooG: Film datasets analysed in Chapter 3 of "Getting (more out of) Graphics"
}

\description{
The film ratings dataset is bigger than 1Mb, so it is provided in this package.
}

\details{
Most of the datasets analysed in the book are available in GmooG and other R packages.
}

\author{
Antony Unwin unwin@math.uni-augsburg.de

}


